rename NewWindowPage -> ReuseWindowConcurrency, move to windowconcur.js potential fix for #156
